Rendering of Views
Why:
So that view templates containing details specific to request can be displayed by applications using the framework.
So that when a template is not specified by the application builder it is searched for and rendered by the framework.

This issues are fixed by:
Making use of the Tilt gem for view rendering.
Writing methods to pass instance variables of controller to the views.
Writing a render and response method in the Controller class.

[Finishes #109951748]