Inclusion of ORM
Why:
So application using gem can model objects.
So application using gem can communicate with database and perform CRUD operations on the DB.
So application using gem  can have access to query methods for database queries.

This issue if fixed by:
Creating an ORM folder.
Creating a BaseModel class with-
1. to_table method to set DB table name.
2. property method to set DB table properties.
3. create_table method to create a table if it doesnt exist.
4. create method to create records/rows in the DB.
5. save and update method save records or update them.
Creating a Queries class that is inherited by the BaseModel with methods for querying the DB for information.